chore(update-timestats.yml): add GitHub Actions workflow to update WakaTime statistics in the Readme file
This commit adds a new file `.github/workflows/update-timestats.yml` which contains a GitHub Actions workflow. The workflow is triggered on a schedule, running every day at 12:00 PM.

The workflow job named "update-readme" is configured to run on the latest version of Ubuntu. It consists of a single step that uses the `anmol098/waka-readme-stats` action. The action is configured with various input parameters to customize the behavior of the generated metrics.

The action retrieves the WakaTime statistics using the provided WakaTime API key and updates the Readme file in the repository. The commit message for the update is set to "Updated dev metrics" and the author of the commit is set to "The_R4V3N" with the email address "contact@oliver-joisten.se".

The generated metrics include information about the user's programming language usage, lines of code, and commit activity. Some metrics are set to be shown in the Readme file, such as projects, updated date, and short information. Other metrics, such as OS, editors, and days of the week, are not shown.