chore(waka_stats.yml): add GitHub Actions workflow for updating WakaTime stats in the README
This commit adds a new file `.github/workflows/waka_stats.yml` which contains a GitHub Actions workflow. The workflow is triggered daily at 12 am UTC using a cron schedule. It runs on the latest version of Ubuntu and consists of a single job named "update-readme".

The job uses the `avinal/Profile-Readme-WakaTime` action to update the WakaTime stats in the README file. It requires the `WAKATIME_API_KEY` secret to be set in the repository's secrets. The action also uses the `GITHUB_TOKEN` provided by GitHub for authentication.

The action is configured with the following parameters:
- `BRANCH` is set to "master" as the default branch.
- `COMMIT_MSG` is set to "Automated Coding Activity Update :alien:" as the commit message for the update.
- `STATS_RANGE` is set to "last_7_days" to fetch the coding activity data for the last 7 days.

This workflow will automatically update the WakaTime stats in the README file, providing insights into the coding activity of the repository.