feat(WakaTime.yml): add WakaTime status update workflow
This commit adds a new workflow file, WakaTime.yml, which sets up a scheduled job to update the WakaTime metrics in the README.md file. The workflow is triggered manually or on a daily schedule at 12 AM UTC (5:30 AM IST).

The job runs on the latest version of Ubuntu and consists of the following steps:
- Uses the "athul/waka-readme" action to update the WakaTime metrics in the README.md file.
- Requires the "GH_TOKEN" and "WAKATIME_API_KEY" secrets to authenticate with GitHub and WakaTime API, respectively.
- Provides optional configuration parameters for customizing the content and appearance of the metrics.
- Supports committing the changes to a specified branch and file, with customizable commit details.

The commit message for the WakaTime metrics update is set to "Updated waka-readme graph with new metrics" (optional) and the commit is attributed to the GitHubActionBot (optional).