chore(latest-wakatime-activity.yml): remove latest-wakatime-activity workflow
The latest-wakatime-activity.yml file has been deleted. This workflow was responsible for updating the README file with the latest Wakatime activity. It was triggered manually or on a daily schedule at 12AM UTC.

The decision to remove this workflow was made to simplify the repository and reduce unnecessary automated processes.