feat(WakaTime.yml): add WakaTime workflow to update README with latest activity
The WakaTime.yml file is added to the .github/workflows directory. This file contains a workflow that is triggered by a manual dispatch or a scheduled cron job. The workflow is named "Latest Wakatime Activity".

The workflow has a single job named "update-readme" that runs on the latest version of Ubuntu. It consists of several steps. The first step uses a custom action from The_R4V3N/waka-readme repository to update the README file of this repository. The action requires the WAKATIME_API_KEY and GH_TOKEN secrets to be set in the repository settings.

The action is configured with the following options:
- SHOW_PROJECTS is set to "False" to hide project information in the README.
- SHOW_LINES_OF_CODE is set to "True" to display lines of code information in the README.
- SHOW_LOC_CHART is set to "False" to hide the lines of code chart in the README.
- SHOW_TITLE is set to false to hide the title in the README.
- LOCALE is set to "en" to use English language for the README.

This workflow will run at every 12AM UTC as specified by the cron schedule. It will update the README with the latest WakaTime activity.