feat(waka-readme.yml): add Waka Readme workflow to update dev metrics in README
The Waka Readme workflow is added to automatically update the dev metrics in the README file. It is triggered manually or scheduled to run daily at 12 AM UTC (5:30 AM IST). The workflow runs on the latest version of Ubuntu and consists of a single step that uses the "The-R4V3N/waka-readme" action. The WAKATIME_API_KEY secret is passed to the action as an input.