chore(Waka.yaml): add Waka Readme workflow to update readme with metrics
This commit adds a new file, `.github/workflows/Waka.yaml`, which contains the configuration for a GitHub Actions workflow. The workflow is triggered on a schedule, running at 12am IST, and also can be manually triggered using the workflow_dispatch event.

The workflow has a single job named "update-readme" that runs on the latest version of Ubuntu. It consists of several steps, the first of which uses a custom action from The-R4V3N repository. This action requires the `WAKATIME_API_KEY` and `GH_TOKEN` secrets to be set in the repository.

The action is configured with various options to control the behavior of the metrics displayed in the readme. These options include hiding the operating system, hiding projects, committing only the changes made by the user, setting the commit username, showing the updated date in a specific format, and showing the lines of code.

This workflow will automatically update the readme with the metrics provided by WakaTime.