chore(Metrics): remove Metrics workflow file
The Metrics workflow file has been deleted. This file was responsible for generating metrics and visualizations for the GitHub user "The-R4V3N". It was scheduled to run every hour and also triggered manually and on each commit to the "master" and "main" branches.

The decision to remove this workflow was made because it is no longer needed or relevant for the project.