feat(Metrics.yaml): add Metrics workflow to generate GitHub metrics
The Metrics.yaml file is added to the .github/workflows directory. This file contains the configuration for a GitHub Actions workflow that generates metrics for the repository. The workflow is triggered on a schedule, running every hour, and can also be manually triggered or triggered on each commit to the "master" or "main" branches.

The workflow uses the lowlighter/metrics action to generate the metrics. It requires a GitHub token with the necessary scopes to access the required data. The token is stored as a secret named METRICS_TOKEN.

The generated metrics include various information about the user The-R4V3N, such as activity, community, repositories, and metadata. The metrics are generated using the "classic" template and are configured to display information for the past 14 days. The metrics also include information about habits, isocalendar, lines of code, people, stars, traffic, and Wakatime integration.

The generated metrics are displayed in various sections and charts, providing insights into the user's activity, habits, code usage, and more.

The metrics are generated using the Europe/Stockholm timezone.

Note: Some of the plugin configurations have additional options and limits set, such as the number of activities, habits, lines of code, people, stars, and Wakatime data to display.