chore(Metrics.yaml): remove Metrics workflow file
The Metrics.yaml file was deleted. This file contained a GitHub Actions workflow that generated metrics for a user's GitHub activity. The workflow was scheduled to run every hour and also triggered manually and on each commit to the "master" and "main" branches. It used the lowlighter/metrics action and required a GitHub token with specific scopes.

The removal of this file was done to remove unnecessary metrics generation and reduce the complexity of the repository's workflows.